tax_env/requirements.txt
streamlit==1.45.1
python-dotenv==1.0.1
langchain==0.2.16
langchain-openai==0.1.23
langchain-community==0.2.16
langchain-text-splitters==0.2.4
langchain-chroma==0.1.3
langchain-pinecone==0.1.3
pinecone-client==5.0.1
openai==1.84.0
python-dotenv==1.0.1